refactor: formatDuration を utils に
